Remove reference to COCO labels file (#84)
In the docker compose file the acap_dl-models volume was mounted in the
pose-estimator container and an environment variable specified the path
to a COCO dataset labels file. This file does not exist and the
environment variable is never used, it thus seems like the extra lines
in the docker compose file are leftovers from the object-detector-python
example.